Remove Colonize Screen
The Colonize Screen didn't provide much in the way of spectacle and
added needless clicking to advancing the turn order. Now the colony
name can be entered on the Colonize Map Overlay removing the need for
the additional screen to handle colonization.

Ideally, the Map Overlay would use a JDK Text Box to take colony name
input but due to the way Overlays sit in the paint cycle I wasn't able
to make this work. This may be addressed in another commit perhaps.

Furthermore, the colonization jingles have not been preserved. This is
because in the little time I spent looking into it I didn't figure out
how it could be done. Again, future work may address this.

(If it seems as if a lot of work was not complete for this commit -
look at the gap in the commits. It has proved to be very draining and
I wish to move on. I can do this because it is a hobby project, after
all.)