Colour neutral star names by habitability
Uncolonised star system names are now coloured based on wherever they
are habitable to the viewing empire. This makes it significantly
easier to see where points of interest are during the expansion part
of the game.

They used to be coloured based on wherever they were in range –
however this is now displayed in the Map Overlay so is redundant.

The colour for habitable/in range star system has also been turned up
from light grey to white to make it a more noticeable.