Remove textures
Map textures made it significantly harder to parse what is happening
on the map and the UI textures, while not terrible, didn't seem to add
much/looked a bit naff. The concept can be revisited in the future but
the current implementation is better removed than kept (as decreed by
me).