Default max bases to 0
It is hard to guess how many bases should be built in the general case
(and if you could, the ability to change it should be removed since it
would be just adding busy work). 0 is probably a better guess than 1
because 0 bases is at least a good guess for the beginning of the
game.