Remove Ship Display Control
Ship Display Control seemed to be buggy and in the fashion that should
by now be typical of this fork - instead of fixing the feature, I
removed it. The feature seems to not be that useful (I haven't found
much use for it having played with the game for a while now).