Simplify GalaxyMapPanel initialisation code
init0() is not a very descriptive name for a method and is only done
as part of instance construction. It makes sense to merge it with the
constructor proper to remove a layer of indirection when reading the
code.

It also doesn't make sense for instance constructor to initialise
static fields (showDistance), especially if it is already initialised
in line.