Remove First Contact Incident
Lowering relations on first contact delays the beginning of the
diplomacy game because you have to wait for the penalty to wear off
before you can start making trade deals and exchange technology. This
doesn't seem to have any impact on the dynamics of the game but
creates timers for the player to keep track of in order to come back
to species they have already made contact. Basically, it adds
cognitive load without adding any "fun" - hence it is removed.