Reduce AI diff
I intend to merge all the AIs into one. To do this I am reducing the
diff between them one bit at a time.

In this case the diff between base and modnar is due to a bug.
"(float) (i/j)" performs integer division (dropping the remainder) and
then rounds. What was probably intended (and what is in base AI) is
"(float) i/j" which converts i to a float and this forces j to be
converted to a float before the division occurs - preserving the
remainder.