Replace modnar AI General with base
Trying to resolve individual differences between base and modnar is
taking a long time and the differences don't seem to be particularly
profound. As such I am taking a more aggressive approach of simply
removing modnar AI classes.

The reason for trying to remove/merge modnar AI into the base AI is
that I would like to rework large chunks of the AI but having three
places to change instead of one is an undesirable maintenance cost
that I don't want to keep paying other the course of the project.