Remove Spiral map shape
The main feature of this map (spiral arms) does not work well with the
mechanics of the game as spawning on the arms severely limits your
early expansion options. This is a balance concern but also makes the
game less interesting to play.

As should be typical by now, instead of trying to change this feature
to work in some way - remove it to make for a more streamlined
experience and simpler codebase.