Remove Year/Turn distinction
Race specific dating systems are ignored and now the turn count is
always referred to as the “year” in the UI. This removes one more
setting - simplifying the code in the process.

Furthermore, thematically, it's hard for me to imagine that the
beginning of an interstellar empire is not the beginning of an age for
a civilisation.