Remove minimum population loss to trigger war
Allowing the player/AI to bomb a colony without automatically starting
a war doesn't seem to change the dynamics of the diplomacy game and
leads to some semi-boring game play to try to use this optimally. As
such this feature is removed, making the Xilmi and base AI closer to
each other.