Unify AI Spy Confession Incident behaviour
The goal is to merge Xilmi and base AI together so as to have less
code to maintain and simpler configurations for the user.

All leaders are now treated as hating spies as, unlike the original
game, spies can now be "removed" and I think less intelligence can
make for a better game (this decision will obviously be reversed if
that doesn't pan out).