Remove Leader methods from Diplomat
Neither of the Diplomat subclasses do anything different other than
passing the call directly to the Leader implementations so this feels
like a needless layer of indirection.

Furthermore, Diplomats shouldn't need to forward calls to Leaders as
that needlessly expands the API surface area. If different Diplomat
implementations need to have different Leaders, different
implementations of Leader should be used instead. This is an unlikely
direction for the project to take as the intent is to merge both
Diplomat implementations but is mentioned for completeness.