Remove Kill Monster Incident
In my mind, the design goal of Space Monsters is to knock leading
empires down a peg, or at least expose a weakness for their enemies.
Giving out a reward for destroying a Space Monster undermines that
goal. As such this incident is removed.

It is debatable if Space Monsters currently achieve this design goal
and there will likely need to be more work done here but this is a
step in the direction I want to take things in.