Remove special Incident methods from Diplomat
All methods do to with noticing special Incidents have been removed
from the Diplomat interface and moved to respective Incident
implementations. This is because, on my reading, the Diplomat should
be responsible for making decisions based on Incidents/state rather
than populating Incidents themselves. All of removed methods did the
same thing across different implementations which also removes some
code duplication.

I think it makes sense for Diplomat to calculate and set the severity
of the Incident in principle but the current architecture doesn't
support that in (what I think is a) clean way so it will be addressed
in later work.