Remove research map size adjustment
The adjustment doesn't quite work because the galaxy is rarely, if
ever, split evenly.

But, more importantly, I don't think making research take the same
amount of time across all possible map configurations is a good goal
as developing research at different rates makes different map sizes
more interesting (to me). This may be taken back later but currently I
am opting for simplifying the code and game mechanics.

Lastly, the base research is cost is set to 40 as this is in line with
hardest difficulty setting in the original game (for the player).