Clean up combat package
An initial attempt to clean up the combat package. i.e. removing
unused methods and fields; making everything that can be private
private; etc.

There is probably a lot more that can be done (and a lot that needs to
be done) but to limit the scope of the change I'm cutting it here.