Refactor TradeRoute
A general clean-up of the code and a change of formatting to something
that I am used to.

I've ummed and ummed about wherever I should stick with the style
currently present in the codebase but after just trying out the
formatting changes I am much happier and more effective. So this will
probably become the norm before (significant) changes to a class are
made.