Remove newColor() from Base
Base has poorly defined responsibilities and cross domain knowledge
which makes it bad design in of itself but descending most classes in
the project from it makes this a particular bad thing from design
perspective. So this is a start on a long journey to cut down Base to
its bare essentials and then figure out what to do with it from there.

In this particular case, there is no reason that I can see to wrap a
call to Color constructor with a nearly identical method signature. So
it is removed and replaced with a direct call to the constructor.