Use trade max profit when displayed with profit
When playing as the Humans it is confusing to see 213BC profits out of
200 BC treaty and not know how soon the maximum is going to reached
(without doing some maths). This happens because the trade level is
used instead of max profits in those places. This is, presumably,
because it was assumed that trade level is max profits but this does
not hold for Humans who have a bonus to trade.

This change replaces trade level with max profits in all places that
seem appropriate.