Remove abs() from Base
As talked about in earlier Base method removal, Base is poorly
designed and this is a continuation of an attempt to cull its
influence on the code.

In this particular case there is no obvious (to me) benefit to warp a
static call to JDK class in a default interface method. As such it is
removed.