Fix a bug when invading transports are wiped out
This was introduced by me in a change where enemy fleets could stop
friendly (to the colony) transports from landing (or at least reduce
their number). The cause of this bug is me missing an early return
when I pulled code to a common helper method - meaning that a check
for zero transport size was effectively missing - resulting in a
divide by zero error.

This change puts the check back in.