Rework CouncilVoteIncident
Previously, no incident was raised if the empire voted for itself in
the election. This undermines the original design goal of the incident
(in my opinion) which is supposed to drive conflict between the two
super powers in the galaxy. This is now changed.

The severity has also been simplified since I don't think the state of
the relationship between two empires should play a factor in this
dynamic.