Stop dead ships from firing
Previously, the AI would try to move and then attack with a ship stack
under their control. Because there is no check to see if the ship
stack has died due to missile fire (or other means) during the move
the AI would attack even if the ships have all already died. This
spawn zero sized missile stacks which clutter up the UI.

Ideally, the AI wouldn't be able to play the game "wrong" but that
would require a significant rework of the way this bit of the code
works and is outside of the scope of what I would like to commit to
right now.

For the time being, simply count dead ship stacks as not able to
attack. This seems to result in desired effect and work across the
different AI flavours that currently exist.