Remove SignTradeIncident
Signing a trade agreement should be an investment to improve relations
- therefore there shouldn't be any immediate kickback for signing a
trade agreement.

This change also introduces DiplomaticReplies class. This is added to
migrate diplomacy related dialogue generation out of the core business
logic of the game/AI. Something that will, no doubt, take many commits
to see to the end.