Rework relations calculations
Previously each incident had an individual timer for how long it took
effect. This meant that if you did two (or three, or four...) bad
things at the same time you would be forgiven in the same amount of
time as if you did one.

This does not feel very satisfying to me because it creates a sense of
the relationship being static. i.e. Things drift back to the starting
point very easily. It also overemphasises the importance of big
diplomatic events vs small continues events, in my opinion.

To this end, the relationship calculations have been completely
reworked. Now each incident contributes to the overall relationship
score and the score, as a whole, drifts towards the base relationship
at a "fixed" rate. This means two bad events are going to be forgiven
in twice the time (roughly) and small reoccurring incidents can build
up to have a large impact.

The relative impact of some incidents has been adjusted to take the
new functionality in mind but the balance of incident severity almost
certainly will need further adjustments. None the less, I feel like
this is already a marked improvement over previous functionality.