Remove Alliance Victory
This win condition is redundant as:

If you are a Galactic Council candidate, your allies can demonstrate
their loyalty by voting for you; if they don't then I don't think
you've won.

If you are not a Galactic Council candidate, you are in a prime
position to win via Council Alliance condition; if can't then there is
still a game to play.