Stop player from aiding multiple times per turn
Currently, if the player aids another empire in the same turn the
existing incident will be adjusted but this will not have any impact
on the relationship between the empires as that is calculated at the
point the incident is first introduced. This is something that I have
broken during the relationship rework.

To resolve this the player can no longer offer aid multiple times per
turn. I think this makes the aid a more significant diplomatic action
and is generally a move in the right direction, in terms of game
mechanics.

The implementation is a bit sloppy/placed in probably not the right
place. This is due to the current code already distributing the
responsibilities weirdly, in my opinion, and I don't want to spend the
time reworking it now. Future rework will probably follow eventually.