Store all incidents as they occur
Because incident impact on the relationship is now only calculated
when it is newly added (and doing otherwise is not straight forward
and, maybe, not desirable) remove the functionality of overwriting
existing incidents. Likewise, "less sever" incidents are no longer
ignored.

This almost certainly introduces new bugs where an incident is going
to be added twice (as this was previously hidden by the overwriting of
the same incidents) but is generally a step in the right direction and
will need to be fixed later.