Refactor receive*Aid() AI methods
These methods implement game mechanics and build string to display to
the user - neither of which should be the responsibility of the AI.
Move the game mechanics logic to DiplomaticEmbassy (this is not a
perfect fit but is an improvement and will do for now). Move the
string building to DiplomaticReplies.