Automate replacement of generic tokens
Initially I had wanted to be explicit in regard to which tokens are
replaced in a diplomatic reply but on further thought I decided to
maintain the "generic token" concept existing in the current code.
This means that changing the tokens in the diplomatic reply string
doesn't necessarily require a code change and will probably come in
handy later.