Remove System alarms for explore tab
The UI already shows what stars are explored and which are within
range so having them flash is redundant/unhelpful.

It is possible we might want to remove the explore tab completely in
the future but for now it is being left due to the labelling it puts
on not owned stars.