Remove tabs
This commit does a bunch of whitespace clean up but the focus is the
eradication of tabs in the code base. A lot of tabs are due to my
changes and if it was up to me the code would be tab indented but the
code as-is is mostly space indented so we are going for consistency.

All tabs should be removed and replaced with four spaces. This should
make it easier to read the code in an environment where tab lengths
are not the default four spaces.