Clean up code
Storing data in a Map as String for a key and Object as value denies
you type safety and allows you to make mistakes with the spelling of
the key without the compiler complaining - both of which can be solved
by simply replacing the lookup with a field. This should be a
relatively uncontroversial code quality improvement.