Remove Diplomat.readyForWar()
This method was not implemented by either of the Diplomat
implementations. While I think I have an idea of what the
implementation was intended to be I instead opted to simply remove the
method. This is done because I want to move to a design where the
majority of the decisions the AI makes are based on relationship value
they have with the other players rather than other factors.