Convert isShip() to isEmpireShip()
The initial goal was to absorb all space monsters and "empire" ships
into one class and, while that is still the goal, that is proving more
difficult than initially assumed. As CombatEmpireShip is now likely to
exist for longer than initially planned and CombatShip will likely be
an abstract super class of CombatEmpireShip and, eventual,
CombatMonster classes (at least for a time) - move previous isShip()
checks to isEmpireShip() checks and convert CombatShip casts to
CombatEmpireShip casts (the last is related because the existing code
assumes that if isShip() is true it is safe to cast to CombatShip) as
this will maintain parity with previous logic and allow for more
subclasses of CombatShip to be created.

For further context: CombatEmpireShip was an attempt to extract
everything "empire" related from CombatShip with the goal of using
CombatShip to represent space monsters. Currently, however, nothing
instantiates CombatShip so previous isShip() checks can be safely
converted to isEmpireShip() check (and likewise with the casts).