Fix Colony Scan Information pip dimensions
The Colony Scan Information panel has little coloured pips next to the
stats which are shown on the tactical view - so that you could
discover what the numbers in the coloured circles mean in the tactical
view. These pips are more oval than circle for the Colony panel (the
Ship panel has circles) and the tactical view pips are circles as
well. To make things consistent the Colony panel pips are now circles
too.