Remove initiativeRank() method
The difference between initiativeRank() and initiative() is confusing
from an API point of view. Effectively, initiative() is a partial
calculation of the initiative that is dependent on different
CombatEntity implementations; while initiativeRank() is what is
expected to be used externally and is the final "initiative score".

The CombatEntity implementation of initiativeRank() was only relevant
to CombatShip though (as only it can teleport or cloak) so move it
there which makes initiativeRank() redundant so it can be removed.

Outside of initiativeRank() nothing uses initiative() so this change
is safe.