Make Xilmi AIDiplomat extend Base AI
The intent is to merge the two different AIs over time to reduce code
to maintain and merge the best parts of the both AIs into one (best
being decided by the project maintainer - me). This is a convenient
way to transition to this goal as it allows me to see what is
different between the two AIs better and adjust code without breaking
functionality or duplicating code.