Remove Xilmi AI council vote logic
The Xilmi AI doesn't base any of the council voting decision around
the relations rating between Empires which goes against the design
purpose of the relations rating - to indicate the disposition of the
AI towards the player. While this makes the AI less like the player
(the player doesn't broadcast her or his opinion about the AIs) it
makes for more satisfying gameplay (in my opinion) by giving the
player the necessary information to make strategic decisions.

As the result of this the Xilmi AI council vote logic is now removed
and the method is inherited from the base AI.

This also allows me to remove some code in the AI General classes as
it is no longer used.