Reinstate ambassador withdrawal to Xilmi AI
Ambassador withdrawal probably shouldn't be something that a player
can decide to enact as it doesn't seem to bring any interesting
decisions and should probably be thought of as part of the game
mechanics. Moving it out of AI classes is probably too much for me to
attempt right now, however. For now simply reinstate the ambassador
withdrawal logic to Xilmi AI as this brings the two AIs closer
together which bring the goal of merging the two AIs ever closer.