Refocus new incident factory methods
The incident factory methods should only be responsible for creating
the incident and not adding them to the embassy. This helps to reduce
how much incidents need to know about the context in which they are
used. This should help to break circular dependencies in the future
and improves code quality.