Ignore flying missiles when ending combat
This special case seems to only impact situations when a Colony is
attacked and the bases are destroyed but the missiles from the bases
are still in flight. This drags out the combat and feels inconsistent
with how the rest of the combat works (ship that is blown up has its
missiles deactivated).

This is now removed.