Remove endOfCombat() from combatIsFinished()
This removes duplicate calls to endOfCombat() as combatIsFinished()
would be called by the UI to check if combat was over (causing
endOfCombat() call in the process) and by the CombatManager when
combat was over. The duplicate calls are a problem because they
duplicate incidents for a given combat encounter if the combat is for
the player and uses the UI.

This was not a problem in the original Remnants of the Precursors
because the duplicate incident would be debounced by the map storage
of incidents (i.e. only one incident for a key was allowed). However,
as we moved to a different model of handling incidents and evaluating
relationship this debouncing no longer happened which exposed the
duplicate calls.