Merge similar diplomatic incidents
In the long run it would be nice if all incidents where just a
parameterisation of the same class, however this is not something
easily achievable in the current design. As such this is going to be
something that will be chipped away a little bit at the time (to
maintain adequate levels of motivation).

The reason for migrating to this design is "better code reuse" as this
would allow me to make changes to all incidents easily (which is
something I want to do currently).