Remove BioweaponIncident
Decreasing your relationship with all empires is only strategically
viable decision when you are in a position to be at war with all
empires (i.e. you have already won). This effectively forbids
bioweapons until the last couple dozen turns which is too restrictive
for a technology that is apparently designed as a viable alternative
to other technologies. To that end, remove the incident to increase
the viability of using bioweapons.

(The unexplored alternative in the above logic is that the incident
could just impact the relationship score too little to bring you into
conflict - but then the incident would effectively have too little
impact to be noticed - so why bother having an incident like that at
all in the first place?)