Make Tech.cat private
In general, all fields should be private so that the access can be
controlled via setters/getters and internal state of a class can be
hidden from external callers.

More specifically, this work is done so that Tech.cat can be removed
from Tech. This is in part because currently there is a circular
dependency between Tech and TechCategory which makes code harder to
understand and update but also there is a plan to change how Tech is
initialised which requires this to be done.