Remove year count from prompts and notices
A lot of the time the year/turn count of the game has no impact on the
decision or notice being presented. Despite this, however, the year
count would take prominent position in pretty much all notices and
prompts. This pollutes the clarity of the information presented to the
player with noise and limits the space that can be used for useful
information. To this end the year count has been removed from all
places it was not deemed useful.

Simply removing the year count, however, left an unseemly gap in the
flow of the layout - so the layout for many prompts and notices has
been adjusted to look better and more consistent across the board.