Correct string value
troopMsg would be used to concatenate with other strings to form user
facing messages. This resulted in "null" being printed to the user.
This is almost certainly a bug and has been replaced with empty string
to simply omit the troop message when there are no troops to message
about.