Remove method from Base interface
Base interface contains a large number of default methods that apply
to different domains which are enabled by static access to various
classes. This breaks single responsibility principle and goes against
object oriented programming in general. As such it is the long term
goal of this project to remove Base. However, this is a monumental
task. To make it more achievable it will be done in hundreds of
smaller changes like this.