Remove MapOverlay.drawSprites()
This is currently used to decide if MapOverlay buttons and
subcomponent need to be painted - however as buttons and subcomponents
are now only painted when the MapOverlay is painted this check is now
redundant - you always want to paint the subcomponent if the
MapOverlay is painted. Removing this simplifies the code and makes it
easier to maintain.