Refactor canOfferTradeTreaty()
canOfferTradeTreaty() is conceptually about wherever it is possible to
make a trade rather than wherever it is desirable to make a trade
(which suggests it probably shouldn't be in the AI classes but we will
cross that bridge when we get to it). To that end, the code that
relates to AI decision making has been pushed up the callstack (it is
used by the AI only in one method). This aligns the code better with
single responsibility principle and simplifies some of the logic.