Remove unity related logic
It is long since has been impossible to be in Unity with another race
due to removal of final war. The removal of this method wasn't 100%
straight forward as it returned true if the provided empire was the
same as "this" empire - however this input was never a possible in any
of the places it was called as contactedEmpires() never returned
"this" empire.