Remove Bengali "translation"
None of the translation has been complete and it is not selectable
from the language menu.

At a later point, when the game text has been changed so that it can
be licenced under GPL, it would be really nice to support Bengali as
it is a very widely spoken language. For now, however, it makes sense
to just remove files that don't add any value to the project.