Remove player colony defence alerts
It is perfectly valid (and even optimal) to defer developing shields
until they are needed for a given colony and even completely skipping
them for colonies that are far too remote from the frontlines. As the
result, highlighting your colonies with "underdeveloped" defences
creates a lot of visual noise - effectively alerting you to something
that probably doesn't need addressing - over and over again. To this
end, remove those alerts.