Stop TrespassingIncident during war
Thematically, it's very strange to me that the majority of my negative
relations with the race I am at war with would come from
"trespassing".

Mechanically, a continues build up of negative relations during a war
encourages wars to the death (because even if peace is signed war is
inevitable because of poor relations). A war to the death does not
make for a very dynamic political game and is therefore undesired. The
removal of trespassing incident during war doesn't fully address this
but is a step in the right direction.