Remove Xilmi AI immediateRefusalToTrade()
The only difference between this and the base implementation is that
Xilmi's AI does not consider the Empire relations in its logic. This
goes against the current direction of development which attempts to
make use of relations rating to inject dynamism into the diplomacy
game and as a meaningful communication tool to the player. To that
end, remove Xilmi AI immediateRefusalToTrade().