Simplify decline reason logic
Brining up an incident as the reason for refusing some diplomatic
dealing is more misleading than informative since the relationship
rating logic rework. This is because no single incident can be said to
be dragging your relationship rating down - instead it is an aggregate
of a lot of incidents that contribute to a poor relationship score. As
such, always return the "generic" refusal only.