Remove immediate trade refusal
While the extra sass from the AI seems nice in theory it rarely comes
into play. Therefore, I am removing a low impact feature to simplify
the code - the main advantage of which is that the trade
refusal/acceptance logic is duplicated in fewer places.

An alternative considered but not taken was to simply not offer trade
if it is guaranteed to be refused. This was rejected though because it
would have duplicated the refusal/acceptance logic somewhere else.