Fix null pointer for monsters
This currently only effects the Guardian. If the CombatManager.system
has not been initialised a null pointer will be thrown when the
Guardian constructor attempts to get the system to decide if the
shield should be active or not due to nebula effect. This is clearly a
circular dependency and should be addressed but is deemed out of scope
for now (though progress to this goal will be made in small steps). To
address for this for the time being simply set system before Guardian
stacks are constructed.

To extend this to other monsters (as the intent is to move them to the
same code as the Guardian soon[tm]) stack initialisation had to be
delegated to CombatManager which in turn meant that monster.alive()
would not return correct result if there was no combat. This is
now addressed by hacking in a new boolean flag. A better solution was
deemed to be out of scope for now.