Make Human trade benefit both players
The functionality change here is that when one player has a trade
bonus this impacts both parties. i.e. When you make a trade deal with
Humans, both you and the Humans get a larger maximum trade profit.
This does not impact the initial minimum profit which Humans still
have an edge in.

This is done to make Human bonus more interactive for the opponent -
hopefully impacting the opponent decision making and creating a more
varied experience from game to game. This has a mixed impact of the
Human power as this improves relations but gives more resources to
their opponents. This seems to be ok during play tests and seems like
a better allocation of power.

The rest of the changes is a refactor that followed from the
functionality change:
* As maxProfit() and level() now returned the same value so they have
been merged.
* Because the relationship between "level" and "profit" is not obvious
(to me) refactor "level" to be about "profit limit" - which is
hopefully more clear.