Reduce max number of opponents to 9
Similar to the recent maximum galaxy size changes - while new
computers can handle more opponents it does not lead to a better
experience. To this end, limit the maximum number of opponents to a
more appropriate limit. The limit/ideal number of opponents might need
further adjustment but this is a step in the right direction.