Remove special Alliance breakup logic
The current implementation puts a lot of pressure on alliance members
to do stuff for their fellow members - so much so that (in my
experience) alliances inevitably break up not long after they are
formed. From gameplay perspective this limits the interesting
possibilities for the diplomatic game. There could be an argument for
reducing the severity of the break up in the logic to address this but
it ultimately tracks the same things relations rating already tracks
(albeit in a more focused and demanding way). To this end the special
Alliance break up logic is removed and alliances will now break up
based on relations rating falling too low. Should alliances prove too
stable the change can be achieved through existing systems (i.e. by
increasing severity of some incidents if in alliance or by increasing
the relationship rating at which the break up is considered).

Removing the special logic simplifies the code and making it easier to
maintain and reason about.