Ancient Derelict no longer gives restricted tech
Previously, it was possible to receive restricted tech from the
Ancient Derelict random event because the code selecting the tech
never checked the restricted status of the tech. This is now fixed.
In practise this only meant that you could get a Death Ray without
fighting the Guardian, however this paves the way to adding more
restricted weapons (like Space Monster weapons) without them leaking
to the players.