Highlight soon to be habitable planets on map
Previously, all uninhabitable planet names where displayed in grey -
however it is useful to know if you will soon be able to colonise a
planet due to research. To this end, planets that will be colonisable
once current research completes now have their names displayed in
yellow.

The specific colour is chosen to not overpower the white names which
are currently colonisable planets and therefore more important and to
fit in with how the same information is displayed in the details panel
of the planet.