Remove redundant Expand Alerts
This information, where applicable, is now displayed on the main map
screen so it is removed from the Systems Expand tab.

The Expand tab is not removed, however, because the additional
labelling it provides can still be used. In the long run this too will
be provided on the main map screen, allowing the Expand tap to be
removed in turn.