Remove factories from Colonies panel
Factories values vary a lot throughout early game to late game making
it hard to eyeball the current status using the factories column. Due
to this the same data is more effectively conveyed by the combination
of capacity and planet size. To this end remove factories from the
panel to reduce cognitive load when parsing the data.