Remove production from Colonies panel
Production values, like factory values, vary a lot throughout early
game to late game making it hard to eyeball the current status using
the production column. Due to this the same data is more effectively
conveyed by the combination of capacity and planet size. To this end
remove production from the panel to reduce cognitive load when parsing
the data.