Remove shields and bases from Colonies panel
Shield and base values are not really informative without knowing
where the colony is located (i.e. far from the frontlines – should be
0; close to active warfront – should be as developed as possible). Due
to this the Colonies panel (which does not display any map
information) is not the right place for this information. To this end
remove shields and bases from the panel to reduce cognitive load when
parsing the data.