Consolidate tabs in Colonies panel
Since each tab now (thanks to the removals) contains relatively little
information they can be merged to provide all the available
information in one place. This should reduce the number of tabbing the
player has to do and reduces the code needed to maintain.

This change also updates the help page which has become out of date
due to the removals.