Remove Base interface
Base is an example of a "God class" except it's an interface.
Removing/marginalising the impact of the Base interface is a goal of
the project as part of a broader goal to improve code quality. To this
end, remove the interface from classes that don't need it. This is by
no means the last step in removing Base from the code base.