Remove Human bonus to start trade profit
The cost of starting a new trade route should encourage the player to
think of the long term diplomatic prospects with a given empire
(though it is arguable wherever it succeeds in this respect
currently). Removing the cost for Humans (ironically) gives them a
shallower diplomatic game to play. As such the bonus is removed from
the trade starting profit (or lack thereof).